requirements.txt
copilotkit==0.1.0
fastapi>=0.111.1

uvicorn==0.29.0
google-cloud-firestore==2.13.1
langchain-google-vertexai==0.0.5
pydantic==2.5.2
python-multipart==0.0.6
# Dev / Testing
flake8
pytest
httpx
python-dotenv
